fix #82: 웹뷰 결과 파싱 실패 버그 수정
ㅇ

+++ Projects/Domain/Sources/Model/KeymeWebModel.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 public struct KeymeWebViewModel: Codable, Equatable {
-    public let matchRate: Int
+    public let matchRate: Float
     public let resultCode: String
     public let testResultId: Int
 }
